fix: Fix BFloat16 dtype error in MIA min_k_plus_plus attack
Add .float() conversion before calling .numpy() on torch.sqrt(sigma)
to fix TypeError when model outputs are in BFloat16 format.

Also add comprehensive tests for MIA attacks:
- Test dtype compatibility (float32, float16, bfloat16)
- Test edge cases (empty tensors, single element)
- Test numerical stability (uniform distributions)

+++ src/evals/metrics/mia/min_k_plus_plus.py
@@ -32,7 +32,7 @@ class MinKPlusPlusAttack(MinKProbAttack):
         sigma = torch.clamp(sigma, min=1e-6)
         scores = (
             target_prob.float().cpu().numpy() - mu.float().cpu().numpy()
-        ) / torch.sqrt(sigma).cpu().numpy()
+        ) / torch.sqrt(sigma).float().cpu().numpy()
 
         # Take bottom k% as the attack score
         num_k = max(1, int(len(scores) * self.k))
